Update docs for accuracy, fix all clippy warnings and formatting
- README: update test count to 289, add enterprise hardening features,
  fix project status description
- CHANGELOG/SECURITY/GOVERNANCE: add missing copyright headers, fix
  GitHub handle and advisory URL
- CONTRIBUTING: update integration test path pattern
- Fix clippy warnings: const fn on builder, u64-to-usize cast direction,
  async fn syntax in client test interceptor
- Apply rustfmt to all files

+++ crates/a2a-server/src/dispatch/jsonrpc.rs
@@ -241,7 +241,7 @@ async fn read_body_limited(body: Incoming, max_size: usize) -> Result<Bytes, Str
     // Check Content-Length header upfront if present.
     let size_hint = <Incoming as hyper::body::Body>::size_hint(&body);
     if let Some(upper) = size_hint.upper() {
-        if upper as usize > max_size {
+        if upper > max_size as u64 {
             return Err(format!(
                 "request body too large: {upper} bytes exceeds {max_size} byte limit"
             ));
